Marketplace API name change
Marketplace terminology was removed from Checkout API Reference, so in order to matches what API reference contains we migrated all the MarketplaceClient stuff into AccountsClient
Decoupled Balances and Transfers from Marketplace/Accounts into his own clients BalancesClient and TransfersClient
