+++ checkout_sdk/balances/balances.py
@@ -1,0 +1,2 @@
+class BalancesQuery:
+    query: str
