+++ tests/accounts/accounts_payout_schedules_integration_test.py
@@ -5,10 +5,10 @@ import os
 import pytest
 
 import checkout_sdk
+from checkout_sdk.accounts.accounts import UpdateScheduleRequest, ScheduleFrequencyWeeklyRequest, DaySchedule, \
+    ScheduleFrequencyDailyRequest, ScheduleFrequencyMonthlyRequest
 from checkout_sdk.common.enums import Currency
 from checkout_sdk.four.oauth_scopes import OAuthScopes
-from checkout_sdk.marketplace.marketplace import UpdateScheduleRequest, ScheduleFrequencyWeeklyRequest, DaySchedule, \
-    ScheduleFrequencyDailyRequest, ScheduleFrequencyMonthlyRequest
 from tests.checkout_test_utils import assert_response
 
 
@@ -30,13 +30,13 @@ def test_should_update_and_retrieve_weekly_payout_schedules(payout_schedules_api
     schedule_request.threshold = 1000
     schedule_request.recurrence = weekly_request
 
-    payout_schedule = payout_schedules_api.marketplace.update_payout_schedule(
+    payout_schedule = payout_schedules_api.accounts.update_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq', Currency.USD, schedule_request)
 
     assert_response(payout_schedule, 'http_metadata')
     assert payout_schedule.http_metadata.status_code == 200
 
-    retrieve_payout_schedule = payout_schedules_api.marketplace.retrieve_payout_schedule(
+    retrieve_payout_schedule = payout_schedules_api.accounts.retrieve_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq')
 
     assert_response(retrieve_payout_schedule, 'http_metadata',
@@ -55,13 +55,13 @@ def test_should_update_and_retrieve_daily_payout_schedules(payout_schedules_api)
     schedule_request.threshold = 1000
     schedule_request.recurrence = daily_request
 
-    payout_schedule = payout_schedules_api.marketplace.update_payout_schedule(
+    payout_schedule = payout_schedules_api.accounts.update_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq', Currency.USD, schedule_request)
 
     assert_response(payout_schedule, 'http_metadata')
     assert payout_schedule.http_metadata.status_code == 200
 
-    retrieve_payout_schedule = payout_schedules_api.marketplace.retrieve_payout_schedule(
+    retrieve_payout_schedule = payout_schedules_api.accounts.retrieve_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq')
 
     assert_response(retrieve_payout_schedule, 'http_metadata',
@@ -80,13 +80,13 @@ def test_should_update_and_retrieve_monthly_payout_schedules(payout_schedules_ap
     schedule_request.threshold = 1000
     schedule_request.recurrence = montly_request
 
-    payout_schedule = payout_schedules_api.marketplace.update_payout_schedule(
+    payout_schedule = payout_schedules_api.accounts.update_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq', Currency.USD, schedule_request)
 
     assert_response(payout_schedule, 'http_metadata')
     assert payout_schedule.http_metadata.status_code == 200
 
-    retrieve_payout_schedule = payout_schedules_api.marketplace.retrieve_payout_schedule(
+    retrieve_payout_schedule = payout_schedules_api.accounts.retrieve_payout_schedule(
         'ent_sdioy6bajpzxyl3utftdp7legq')
 
     assert_response(retrieve_payout_schedule, 'http_metadata',
